requirements.txt
ctransformers  
langchain
streamlit
Pillow